v2_rebuild/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0.0
alembic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary
aiosqlite
python-dotenv
stripe
requests
pytz
